+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 ### Changed
 
 - Minimal python requirement updated to 3.12
+- `SSHConfigEntry.node` renamed to `SSHConfigEntry.hostname`
 
 ### Fixed